pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cryptography>=3.1",
    "langchain-community>=0.3.16",
    "langchain-ollama>=0.2.2",
    "langchain-openai>=0.3.2",
    "langchain_chroma",
    "langgraph>=0.2.68",
    "langsmith>=0.3.2",
    "tqdm>=4.67.1",
]
